% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-testing.R
\name{bootstrap_compare}
\alias{bootstrap_compare}
\title{Bootstrap Comparison of Search Strategies}
\usage{
bootstrap_compare(
  strategy1_results,
  strategy2_results,
  gold_standard,
  n_bootstrap = 1000
)
}
\arguments{
\item{strategy1_results}{Results from first strategy}

\item{strategy2_results}{Results from second strategy}

\item{gold_standard}{Vector of relevant article IDs}

\item{n_bootstrap}{Number of bootstrap samples}
}
\value{
Bootstrap comparison results
}
\description{
Bootstrap Comparison of Search Strategies
}
