% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_cosine}
\alias{calc_cosine}
\title{Calculate Cosine Similarity}
\usage{
calc_cosine(text1, text2)
}
\arguments{
\item{text1}{First text string}

\item{text2}{Second text string}
}
\value{
Cosine similarity score
}
\description{
Calculate Cosine Similarity
}
