% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance-metrics.R
\name{calc_efficiency}
\alias{calc_efficiency}
\title{Calculate Search Efficiency Metrics}
\usage{
calc_efficiency(search_time, results_count, relevant_count)
}
\arguments{
\item{search_time}{Time taken to execute search (in seconds)}

\item{results_count}{Number of results retrieved}

\item{relevant_count}{Number of relevant results}
}
\description{
Calculate Search Efficiency Metrics
}
