% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_jaccard}
\alias{calc_jaccard}
\title{Calculate Jaccard Similarity}
\usage{
calc_jaccard(text1, text2)
}
\arguments{
\item{text1}{First text string}

\item{text2}{Second text string}
}
\value{
Jaccard similarity score
}
\description{
Calculate Jaccard Similarity
}
