% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{calc_search_stats}
\alias{calc_search_stats}
\title{Calculate Search Result Statistics}
\usage{
calc_search_stats(search_results)
}
\arguments{
\item{search_results}{Data frame with search results}
}
\value{
List of summary statistics
}
\description{
Calculate Search Result Statistics
}
