% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{calc_silhouette}
\alias{calc_silhouette}
\title{Calculate Silhouette Score for Clustering Quality}
\usage{
calc_silhouette(similarity_matrix, cluster_assignments)
}
\arguments{
\item{similarity_matrix}{Similarity matrix between terms}

\item{cluster_assignments}{Cluster assignment vector}
}
\value{
Average silhouette score
}
\description{
Calculate Silhouette Score for Clustering Quality
}
