% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_col_names}
\alias{clean_col_names}
\title{Clean Column Names}
\usage{
clean_col_names(names)
}
\arguments{
\item{names}{Character vector of column names}
}
\value{
Cleaned column names
}
\description{
Clean Column Names
}
