% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{clean_text}
\alias{clean_text}
\title{Clean Text Fields}
\usage{
clean_text(text)
}
\arguments{
\item{text}{Character vector to clean}
}
\value{
Cleaned character vector
}
\description{
Clean Text Fields
}
