% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{cluster_terms}
\alias{cluster_terms}
\title{Advanced Term Clustering for Strategy Development}
\usage{
cluster_terms(terms, n_clusters = 3, method = "kmeans")
}
\arguments{
\item{terms}{Character vector of terms to cluster}

\item{n_clusters}{Number of clusters to create}

\item{method}{Clustering method ("kmeans", "hierarchical")}
}
\value{
List with cluster assignments and recommendations
}
\description{
Advanced Term Clustering for Strategy Development
}
