% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{create_data_dictionary}
\alias{create_data_dictionary}
\title{Create Data Dictionary}
\usage{
create_data_dictionary(file_path, search_results)
}
\arguments{
\item{file_path}{Output file path}

\item{search_results}{Search results data}
}
\description{
Create Data Dictionary
}
