% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{create_package_manifest}
\alias{create_package_manifest}
\title{Create Package Manifest}
\usage{
create_package_manifest(package_dir)
}
\arguments{
\item{package_dir}{Package directory}
}
\description{
Create Package Manifest
}
