% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{create_package_readme}
\alias{create_package_readme}
\title{Create Package README}
\usage{
create_package_readme(package_dir, search_results, analysis_results)
}
\arguments{
\item{package_dir}{Package directory}

\item{search_results}{Search results data}

\item{analysis_results}{Analysis results}
}
\description{
Create Package README
}
