% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{create_prisma}
\alias{create_prisma}
\title{Create PRISMA Flow Diagram with Proper Spacing and Text Enclosure}
\usage{
create_prisma(flow_data)
}
\arguments{
\item{flow_data}{List containing PRISMA flow numbers}
}
\value{
ggplot object
}
\description{
Create PRISMA Flow Diagram with Proper Spacing and Text Enclosure
}
