% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_progress_bar}
\alias{create_progress_bar}
\title{Create Progress Bar for Long Operations}
\usage{
create_progress_bar(total, format = "[:bar] :percent :elapsed")
}
\arguments{
\item{total}{Total number of iterations}

\item{format}{Progress bar format string}
}
\value{
Progress bar object
}
\description{
Create Progress Bar for Long Operations
}
