% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_strategy}
\alias{create_strategy}
\title{Create Default Search Strategy Template}
\usage{
create_strategy(terms, databases, date_range = NULL, filters = NULL)
}
\arguments{
\item{terms}{Character vector of search terms}

\item{databases}{Character vector of databases}

\item{date_range}{Date vector of length 2 (start, end)}

\item{filters}{List of additional filters}
}
\value{
Search strategy list
}
\description{
Create Default Search Strategy Template
}
