% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{create_strategy_variants}
\alias{create_strategy_variants}
\title{Create Strategy Variants for Optimization}
\usage{
create_strategy_variants(original_strategy, suggestions, method = "greedy")
}
\arguments{
\item{original_strategy}{Original search strategy}

\item{suggestions}{Term suggestions from ML engine}

\item{method}{Optimization method}
}
\value{
List of strategy variants
}
\description{
Create Strategy Variants for Optimization
}
