% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_summary}
\alias{create_summary}
\title{Create Summary Statistics Table}
\usage{
create_summary(data, numeric_vars = NULL, categorical_vars = NULL)
}
\arguments{
\item{data}{Data frame}

\item{numeric_vars}{Character vector of numeric variable names}

\item{categorical_vars}{Character vector of categorical variable names}
}
\value{
Summary statistics data frame
}
\description{
Create Summary Statistics Table
}
