% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{detect_doi_dupes}
\alias{detect_doi_dupes}
\title{Detect DOI-based Duplicates}
\usage{
detect_doi_dupes(results)
}
\arguments{
\item{results}{Data frame with search results}
}
\value{
Data frame with DOI duplicates marked
}
\description{
Detect DOI-based Duplicates
}
