% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{detect_exact_dupes}
\alias{detect_exact_dupes}
\title{Detect Exact Duplicates}
\usage{
detect_exact_dupes(results)
}
\arguments{
\item{results}{Data frame with search results}
}
\value{
Data frame with exact duplicates marked
}
\description{
Detect Exact Duplicates
}
