% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{export_metrics_csv}
\alias{export_metrics_csv}
\title{Export Metrics to CSV}
\usage{
export_metrics_csv(metrics, file_path)
}
\arguments{
\item{metrics}{List of calculated metrics}

\item{file_path}{Output file path}
}
\value{
File path of created file
}
\description{
Export Metrics to CSV
}
