% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{export_to_ris}
\alias{export_to_ris}
\title{Export to RIS Format}
\usage{
export_to_ris(search_results, file_path)
}
\arguments{
\item{search_results}{Data frame with search results}

\item{file_path}{Output file path}
}
\value{
File path of created file
}
\description{
Export to RIS Format
}
