% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{export_validation_xlsx}
\alias{export_validation_xlsx}
\title{Export Validation Results to Excel}
\usage{
export_validation_xlsx(validation_results, file_path)
}
\arguments{
\item{validation_results}{Validation results}

\item{file_path}{Output file path}
}
\value{
File path of created file
}
\description{
Export Validation Results to Excel
}
