% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_numbers}
\alias{format_numbers}
\title{Format Numbers for Display}
\usage{
format_numbers(x, digits = 3, percent = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{digits}{Number of decimal places}

\item{percent}{Logical, whether to format as percentage}
}
\value{
Formatted character vector
}
\description{
Format Numbers for Display
}
