% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{genetic_optimize}
\alias{genetic_optimize}
\title{Genetic Algorithm Optimization (Simplified)}
\usage{
genetic_optimize(original_strategy, suggestions)
}
\arguments{
\item{original_strategy}{Original strategy}

\item{suggestions}{Term suggestions}
}
\value{
List of optimized variants
}
\description{
Genetic Algorithm Optimization (Simplified)
}
