% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pkg_versions}
\alias{get_pkg_versions}
\title{Extract Package Version Information}
\usage{
get_pkg_versions(
  packages = c("searchAnalyzeR", "ggplot2", "lubridate", "openxlsx")
)
}
\arguments{
\item{packages}{Character vector of package names}
}
\value{
Data frame with package version information
}
\description{
Extract Package Version Information
}
