% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{md_to_html}
\alias{md_to_html}
\title{Simple Markdown to HTML Converter}
\usage{
md_to_html(markdown_lines)
}
\arguments{
\item{markdown_lines}{Character vector of markdown lines}
}
\value{
Character vector of HTML lines
}
\description{
Simple Markdown to HTML Converter
}
