% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory_management.R
\name{opt_df}
\alias{opt_df}
\title{Memory-Efficient Data Frame}
\usage{
opt_df(df, compress_strings = FALSE, verbose = TRUE)
}
\arguments{
\item{df}{Data frame to optimize}

\item{compress_strings}{Logical, whether to convert character columns to factors}

\item{verbose}{Logical, whether to print memory savings information}
}
\value{
Memory-efficient version of the input data frame
}
\description{
Converts a data frame to a memory-efficient format by optimizing column types.
}
