% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_db_performance}
\alias{plot_db_performance}
\title{Create Database Performance Comparison}
\usage{
plot_db_performance(results_by_database, gold_standard = NULL)
}
\arguments{
\item{results_by_database}{List of result sets by database}

\item{gold_standard}{Vector of relevant article IDs}
}
\value{
ggplot object
}
\description{
Create Database Performance Comparison
}
