% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_keyword_eff}
\alias{plot_keyword_eff}
\title{Create Keyword Effectiveness Analysis Plot}
\usage{
plot_keyword_eff(search_results, search_terms, gold_standard = NULL)
}
\arguments{
\item{search_results}{Data frame with search results}

\item{search_terms}{Vector of search terms}

\item{gold_standard}{Vector of relevant article IDs}
}
\value{
ggplot object
}
\description{
Create Keyword Effectiveness Analysis Plot
}
