% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_pr_curve}
\alias{plot_pr_curve}
\title{Create Precision-Recall Curve}
\usage{
plot_pr_curve(retrieved, relevant, thresholds = seq(0, 1, 0.05))
}
\arguments{
\item{retrieved}{Vector of retrieved article IDs}

\item{relevant}{Vector of relevant article IDs}

\item{thresholds}{Vector of threshold values}
}
\value{
ggplot object
}
\description{
Create Precision-Recall Curve
}
