% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{Create Sensitivity Analysis Heatmap}
\usage{
plot_sensitivity(sensitivity_results)
}
\arguments{
\item{sensitivity_results}{Results from sensitivity analysis}
}
\value{
ggplot object
}
\description{
Create Sensitivity Analysis Heatmap
}
