% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization-functions.R
\name{plot_temporal}
\alias{plot_temporal}
\title{Create Temporal Coverage Plot}
\usage{
plot_temporal(search_results, gold_standard = NULL)
}
\arguments{
\item{search_results}{Data frame with search results including date column}

\item{gold_standard}{Vector of relevant article IDs}
}
\value{
ggplot object
}
\description{
Create Temporal Coverage Plot
}
