% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{plot_term_suggestions}
\alias{plot_term_suggestions}
\title{Visualize Term Suggestion Results}
\usage{
plot_term_suggestions(suggestion_results, evaluation_results = NULL)
}
\arguments{
\item{suggestion_results}{Results from MLTermEngine}

\item{evaluation_results}{Optional evaluation results}
}
\value{
ggplot object
}
\description{
Visualize Term Suggestion Results
}
