% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.term_effectiveness}
\alias{print.term_effectiveness}
\title{Print Method for term_effectiveness Objects}
\usage{
\method{print}{term_effectiveness}(x, ...)
}
\arguments{
\item{x}{A term_effectiveness object}

\item{...}{Further arguments passed to or from other methods}
}
\value{
Invisibly returns the input object
}
\description{
Print Method for term_effectiveness Objects
}
