% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{sa_optimize}
\alias{sa_optimize}
\title{Simulated Annealing Optimization (Simplified)}
\usage{
sa_optimize(original_strategy, suggestions)
}
\arguments{
\item{original_strategy}{Original strategy}

\item{suggestions}{Term suggestions}
}
\value{
List of optimized variants
}
\description{
Simulated Annealing Optimization (Simplified)
}
