% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_divide}
\alias{safe_divide}
\title{Safe Division Function}
\usage{
safe_divide(numerator, denominator, default_value = 0)
}
\arguments{
\item{numerator}{Numerator value}

\item{denominator}{Denominator value}

\item{default_value}{Value to return if denominator is 0}
}
\value{
Division result or default value
}
\description{
Safe Division Function
}
