% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{safe_list_to_df}
\alias{safe_list_to_df}
\title{Convert List to Data Frame Safely}
\usage{
safe_list_to_df(x)
}
\arguments{
\item{x}{List to convert}
}
\value{
Data frame or NULL if conversion fails
}
\description{
Convert List to Data Frame Safely
}
