% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_suggestion_engine.r
\name{sim_strategy_search}
\alias{sim_strategy_search}
\title{Simulate Strategy Search on Corpus}
\usage{
sim_strategy_search(strategy, corpus)
}
\arguments{
\item{strategy}{Search strategy object}

\item{corpus}{Validation corpus}
}
\value{
Vector of retrieved article IDs
}
\description{
Simulate Strategy Search on Corpus
}
