% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{standardize_date}
\alias{standardize_date}
\title{Standardize Date Formats}
\usage{
standardize_date(dates)
}
\arguments{
\item{dates}{Character or Date vector}
}
\value{
Date vector
}
\description{
Standardize Date Formats
}
