% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{std_embase_results}
\alias{std_embase_results}
\title{Standardize Embase Results}
\usage{
std_embase_results(results)
}
\arguments{
\item{results}{Data frame with Embase results}
}
\value{
Standardized data frame
}
\description{
Standardize Embase Results
}
