% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{std_generic_results}
\alias{std_generic_results}
\title{Standardize Generic Results}
\usage{
std_generic_results(results)
}
\arguments{
\item{results}{Data frame with generic results}
}
\value{
Standardized data frame
}
\description{
Standardize Generic Results
}
