% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{std_pubmed_results}
\alias{std_pubmed_results}
\title{Standardize PubMed Results}
\usage{
std_pubmed_results(results)
}
\arguments{
\item{results}{Data frame with PubMed results}
}
\value{
Standardized data frame
}
\description{
Standardize PubMed Results
}
