% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{std_scopus_results}
\alias{std_scopus_results}
\title{Standardize Scopus Results}
\usage{
std_scopus_results(results)
}
\arguments{
\item{results}{Data frame with Scopus results}
}
\value{
Standardized data frame
}
\description{
Standardize Scopus Results
}
