% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{std_wos_results}
\alias{std_wos_results}
\title{Standardize Web of Science Results}
\usage{
std_wos_results(results)
}
\arguments{
\item{results}{Data frame with Web of Science results}
}
\value{
Standardized data frame
}
\description{
Standardize Web of Science Results
}
