% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_date_range}
\alias{validate_date_range}
\title{Validate Date Range}
\usage{
validate_date_range(date_range, allow_future = TRUE)
}
\arguments{
\item{date_range}{Date vector of length 2}

\item{allow_future}{Logical, whether future dates are allowed}
}
\value{
Logical indicating if valid
}
\description{
Validate Date Range
}
