% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_strategy}
\alias{validate_strategy}
\title{Validate Search Strategy Object}
\usage{
validate_strategy(search_strategy)
}
\arguments{
\item{search_strategy}{Search strategy object to validate}
}
\value{
Logical indicating if valid, with warnings for issues
}
\description{
Validate Search Strategy Object
}
