% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export-utilities.R
\name{export_metrics}
\alias{export_metrics}
\title{Export Analysis Metrics}
\usage{
export_metrics(metrics, file_path, format = "xlsx")
}
\arguments{
\item{metrics}{List of calculated metrics}

\item{file_path}{Output file path}

\item{format}{Export format ("csv", "xlsx", "json")}
}
\value{
File path of created file
}
\description{
Export Analysis Metrics
}
